METAMAC-1985 - Añadir constraints a los DSD para simplificar la creación y reutilización de los mismos >> Añadido mensaje de error si no se encuentran ck para versionar
git-svn-id: http://svn.arte-consultores.com/svn/arte/istac/metamac/metamac-statistical-resources/trunk/metamac-statistical-resources@53243 05f23f4b-1f0b-0410-ba1b-ad9509270e59

+++ metamac-statistical-resources-core/src/main/java/org/siemac/metamac/statistical/resources/core/facade/serviceimpl/StatisticalResourcesServiceFacadeImpl.java
@@ -2004,6 +2004,12 @@ public class StatisticalResourcesServiceFacadeImpl extends StatisticalResourcesS
         ConstraintsSecurityUtils.canDeleteContentConstraint(ctx, datasetVersion.getSiemacMetadataStatisticalResource().getStatisticalOperation().getCode(), datasetVersion
                 .getLifeCycleStatisticalResource().getProcStatus());
 
+        if (contentConstraint.isIsFinal()) {
+            // Can not delete final constraint
+            throw MetamacExceptionBuilder.builder()
+                    .withPrincipalException(new MetamacExceptionItem(ServiceExceptionType.CONSTRAINTS_DELETE_FINAL, contentConstraint.getConstraintAttachment().getUrn())).build();
+        }
+
         // Delete
         constraintsService.deleteContentConstraint(ctx, urn);
     }
@@ -2017,6 +2023,12 @@ public class StatisticalResourcesServiceFacadeImpl extends StatisticalResourcesS
         DatasetVersion datasetVersion = obtainDatasetVersionFromContentConstraint(ctx, contentConstraint); // Dataset for extract operation
         ConstraintsSecurityUtils.canSaveForContentConstraint(ctx, datasetVersion.getSiemacMetadataStatisticalResource().getStatisticalOperation().getCode(), datasetVersion
                 .getLifeCycleStatisticalResource().getProcStatus());
+
+        if (contentConstraint.isIsFinal()) {
+            // Can not update final constraint
+            throw MetamacExceptionBuilder.builder()
+                    .withPrincipalException(new MetamacExceptionItem(ServiceExceptionType.CONSTRAINTS_UPDATE_FINAL, contentConstraint.getConstraintAttachment().getUrn())).build();
+        }
 
         // Transform
         RegionReference regionReference = constraintDto2RestMapper.toRegionReference(regionValueDto);
@@ -2037,6 +2049,12 @@ public class StatisticalResourcesServiceFacadeImpl extends StatisticalResourcesS
         DatasetVersion datasetVersion = obtainDatasetVersionFromContentConstraint(ctx, contentConstraint); // Dataset for extract operation
         ConstraintsSecurityUtils.canDeleteRegion(ctx, datasetVersion.getSiemacMetadataStatisticalResource().getStatisticalOperation().getCode(), datasetVersion.getLifeCycleStatisticalResource()
                 .getProcStatus());
+
+        if (contentConstraint.isIsFinal()) {
+            // Can not update final constraint
+            throw MetamacExceptionBuilder.builder()
+                    .withPrincipalException(new MetamacExceptionItem(ServiceExceptionType.CONSTRAINTS_UPDATE_FINAL, contentConstraint.getConstraintAttachment().getUrn())).build();
+        }
 
         // Delete
         constraintsService.deleteRegion(ctx, contentConstraintUrn, regionCode);
